refactor: move graph component into dedicated file
